requirements.txt
sqlite3
streamlit
python-dotenv
hashlib
Groq
